Homework #6/src/main/java/ru/otus/dlyubanevich/dao/AuthorRepositoryJpa.java
package ru.otus.dlyubanevich.dao;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Repository;
import ru.otus.dlyubanevich.domain.Author;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;
import javax.persistence.TypedQuery;
import java.util.List;
import java.util.Optional;

@Repository
@RequiredArgsConstructor
public class AuthorRepositoryJpa implements AuthorRepository {

    @PersistenceContext
    private final EntityManager entityManager;

    @Override
    public Author save(Author author) {
        if (author.getId() == 0){
            entityManager.persist(author);
            return author;
        }else{
            return entityManager.merge(author);
        }
    }

    @Override
    public List<Author> findByName(String firstName, String lastName) {
        TypedQuery<Author> query = entityManager.createQuery(
                "select a from Author a where a.firstName = :firstName and a.lastName = :lastName",
                Author.class
        );
        query.setParameter("firstName", firstName);
        query.setParameter("lastName", lastName);
        return query.getResultList();
    }

    @Override
    public List<Author> getAll() {
        TypedQuery<Author> query = entityManager.createQuery("select a from Author a", Author.class);
        return query.getResultList();
    }

    @Override
    public Optional<Author> findById(long id) {
        return Optional.ofNullable(entityManager.find(Author.class, id));
    }
}
